:boom: Introducing breaking changes.使用mybatis plus的雪花算法主键策略作为默认策略

+++ pig-common/pig-common-core/src/main/java/com/pig4cloud/pig/common/core/constant/CommonConstants.java
@@ -40,7 +40,7 @@ public interface CommonConstants {
 	/**
 	 * 菜单树根节点
 	 */
-	Integer MENU_TREE_ROOT_ID = -1;
+	Long MENU_TREE_ROOT_ID = -1L;
 
 	/**
 	 * 菜单
