+++ pig-upms/pig-upms-api/src/main/java/com/pig4cloud/pig/admin/api/dto/UserInfo.java
@@ -46,7 +46,7 @@ public class UserInfo implements Serializable {
 	/**
 	 * 角色集合
 	 */
-	private Integer[] roles;
+	private Long[] roles;
 
 	/**
 	 * 角色集合
