+++ pig-upms/pig-upms-api/src/main/java/com/pig4cloud/pig/admin/api/vo/UserInfoVO.java
@@ -44,6 +44,6 @@ public class UserInfoVO implements Serializable {
 	/**
 	 * 角色集合
 	 */
-	private Integer[] roles;
+	private Long[] roles;
 
 }
