+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/controller/DictController.java
@@ -59,7 +59,7 @@ public class DictController {
 	 * @return 字典信息
 	 */
 	@GetMapping("/{id}")
-	public R getById(@PathVariable Integer id) {
+	public R getById(@PathVariable Long id) {
 		return R.ok(sysDictService.getById(id));
 	}
 
@@ -104,7 +104,7 @@ public class DictController {
 	@SysLog("删除字典")
 	@DeleteMapping("/{id}")
 	@PreAuthorize("@pms.hasPermission('sys_dict_del')")
-	public R removeById(@PathVariable Integer id) {
+	public R removeById(@PathVariable Long id) {
 		sysDictService.removeDict(id);
 		return R.ok();
 	}
@@ -139,7 +139,7 @@ public class DictController {
 	 * @return R
 	 */
 	@GetMapping("/item/{id}")
-	public R getDictItemById(@PathVariable("id") Integer id) {
+	public R getDictItemById(@PathVariable("id") Long id) {
 		return R.ok(sysDictItemService.getById(id));
 	}
 
@@ -174,7 +174,7 @@ public class DictController {
 	 */
 	@SysLog("删除字典项")
 	@DeleteMapping("/item/{id}")
-	public R removeDictItemById(@PathVariable Integer id) {
+	public R removeDictItemById(@PathVariable Long id) {
 		sysDictItemService.removeDictItem(id);
 		return R.ok();
 	}
