+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/controller/RoleController.java
@@ -51,7 +51,7 @@ public class RoleController {
 	 * @return 角色信息
 	 */
 	@GetMapping("/{id}")
-	public R getById(@PathVariable Integer id) {
+	public R getById(@PathVariable Long id) {
 		return R.ok(sysRoleService.getById(id));
 	}
 
@@ -87,7 +87,7 @@ public class RoleController {
 	@SysLog("删除角色")
 	@DeleteMapping("/{id}")
 	@PreAuthorize("@pms.hasPermission('sys_role_del')")
-	public R removeById(@PathVariable Integer id) {
+	public R removeById(@PathVariable Long id) {
 		return R.ok(sysRoleService.removeRoleById(id));
 	}
