+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/mapper/SysDeptRelationMapper.java
@@ -41,7 +41,7 @@ public interface SysDeptRelationMapper extends BaseMapper<SysDeptRelation> {
 	 * 删除部门节点关系,同时删除所有关联此部门子节点的部门关系
 	 * @param id 待删除的部门节点ID
 	 */
-	void deleteDeptRelationsById(Integer id);
+	void deleteDeptRelationsById(Long id);
 
 	/**
 	 * 新增部门节点关系
