+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/mapper/SysUserMapper.java
@@ -58,7 +58,7 @@ public interface SysUserMapper extends BaseMapper<SysUser> {
 	 * @param id 用户ID
 	 * @return userVo
 	 */
-	UserVO getUserVoById(Integer id);
+	UserVO getUserVoById(Long id);
 
 	/**
 	 * 查询用户列表
