+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/mapper/SysUserRoleMapper.java
@@ -39,6 +39,6 @@ public interface SysUserRoleMapper extends BaseMapper<SysUserRole> {
 	 * @author 寻欢·李
 	 * @date 2017年12月7日 16:31:38
 	 */
-	Boolean deleteByUserId(@Param("userId") Integer userId);
+	Boolean deleteByUserId(@Param("userId") Long userId);
 
 }
