+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/service/SysRoleMenuService.java
@@ -36,6 +36,6 @@ public interface SysRoleMenuService extends IService<SysRoleMenu> {
 	 * @param menuIds 菜单ID拼成的字符串，每个id之间根据逗号分隔
 	 * @return
 	 */
-	Boolean saveRoleMenus(String role, Integer roleId, String menuIds);
+	Boolean saveRoleMenus(String role, Long roleId, String menuIds);
 
 }
