+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/service/SysRoleService.java
@@ -34,6 +34,6 @@ public interface SysRoleService extends IService<SysRole> {
 	 * @param id
 	 * @return
 	 */
-	Boolean removeRoleById(Integer id);
+	Boolean removeRoleById(Long id);
 
 }
