+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/service/impl/SysMenuServiceImpl.java
@@ -62,7 +62,7 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 
 	@Override
 	@Cacheable(value = CacheConstants.MENU_DETAILS, key = "#roleId  + '_menu'", unless = "#result == null")
-	public Set<SysMenu> findMenuByRoleId(Integer roleId) {
+	public Set<SysMenu> findMenuByRoleId(Long roleId) {
 		return baseMapper.listMenusByRoleId(roleId);
 	}
 
@@ -74,7 +74,7 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 	@Override
 	@Transactional(rollbackFor = Exception.class)
 	@CacheEvict(value = CacheConstants.MENU_DETAILS, allEntries = true)
-	public Boolean removeMenuById(Integer id) {
+	public Boolean removeMenuById(Long id) {
 		// 查询父节点为当前节点的节点
 		List<SysMenu> menuList = this.list(Wrappers.<SysMenu>query().lambda().eq(SysMenu::getParentId, id));
 
@@ -98,18 +98,18 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 	 * @return
 	 */
 	@Override
-	public List<Tree<Integer>> treeMenu(boolean lazy, Integer parentId) {
+	public List<Tree<Long>> treeMenu(boolean lazy, Long parentId) {
 		if (!lazy) {
-			List<TreeNode<Integer>> collect = baseMapper
+			List<TreeNode<Long>> collect = baseMapper
 					.selectList(Wrappers.<SysMenu>lambdaQuery().orderByAsc(SysMenu::getSort)).stream()
 					.map(getNodeFunction()).collect(Collectors.toList());
 
 			return TreeUtil.build(collect, CommonConstants.MENU_TREE_ROOT_ID);
 		}
 
-		Integer parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
+		Long parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
 
-		List<TreeNode<Integer>> collect = baseMapper
+		List<TreeNode<Long>> collect = baseMapper
 				.selectList(
 						Wrappers.<SysMenu>lambdaQuery().eq(SysMenu::getParentId, parent).orderByAsc(SysMenu::getSort))
 				.stream().map(getNodeFunction()).collect(Collectors.toList());
@@ -124,18 +124,18 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
 	 * @return
 	 */
 	@Override
-	public List<Tree<Integer>> filterMenu(Set<SysMenu> all, Integer parentId) {
-		List<TreeNode<Integer>> collect = all.stream()
+	public List<Tree<Long>> filterMenu(Set<SysMenu> all, Long parentId) {
+		List<TreeNode<Long>> collect = all.stream()
 				.filter(menu -> MenuTypeEnum.LEFT_MENU.getType().equals(menu.getType()))
 				.filter(menu -> StrUtil.isNotBlank(menu.getPath())).map(getNodeFunction()).collect(Collectors.toList());
-		Integer parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
+		Long parent = parentId == null ? CommonConstants.MENU_TREE_ROOT_ID : parentId;
 		return TreeUtil.build(collect, parent);
 	}
 
 	@NotNull
-	private Function<SysMenu, TreeNode<Integer>> getNodeFunction() {
+	private Function<SysMenu, TreeNode<Long>> getNodeFunction() {
 		return menu -> {
-			TreeNode<Integer> node = new TreeNode<>();
+			TreeNode<Long> node = new TreeNode<>();
 			node.setId(menu.getMenuId());
 			node.setName(menu.getName());
 			node.setParentId(menu.getParentId());
